[Weekly] Initial commit for weekly review (fix #81)
Improve code organization and introduce routing.

- Removed unused dependencies from package.json.
- Added react-router-dom for routing capabilities.
- Created new Home and Weekly components for better navigation.
- Introduced Layout component to manage page structure.
- Updated fetchJiraItems to support date range filtering.
