+++ package.json
@@ -37,7 +37,6 @@
     "date-fns": "^2.30.0"
   },
   "devDependencies": {
-    "@types/chrome": "^0.0.3.0",
     "@types/node": "^20.16.5",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
