+++ pages/new-tab/package.json
@@ -47,6 +47,7 @@
     "openai": "^4.58.2",
     "react-hook-form": "^7.53.2",
     "react-markdown": "^9.0.1",
+    "react-router-dom": "^7.0.1",
     "react-select": "^5.8.1",
     "remark-gfm": "^4.0.0",
     "zod": "^3.23.8"
@@ -56,6 +57,7 @@
     "@extension/tailwindcss-config": "workspace:*",
     "@extension/tsconfig": "workspace:*",
     "@extension/vite-config": "workspace:*",
+    "@types/chrome": "^0.0.246",
     "@types/lodash": "^4.17.7",
     "@types/react-syntax-highlighter": "^15.5.13",
     "cross-env": "^7.0.3",
